requirements.txt
numba
numpy
tqdm
tensorboard
torch-tb-profiler
torchsummary
gymnasium[all]
